feat: 지수 성과 랭킹 조회
- 일부 파일명 변경

+++ src/main/java/com/kueennevercry/findex/dto/response/IndexPerformanceDto.java
@@ -1,13 +1,15 @@
 package com.kueennevercry.findex.dto.response;
 
+import java.math.BigDecimal;
+
 public record IndexPerformanceDto(
     Long indexInfoId,
     String indexClassification,
     String indexName,
     float versus,
     float fluctuationRate,
-    float currentPrice,
-    float beforePrice
+    BigDecimal currentPrice,
+    BigDecimal beforePrice
 ) {
 
 }
